Refactor packages to consider in-/out-of-proc worker
