+++ src/Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Core.Extensions/SwaggerUIExtensions.cs
@@ -5,7 +5,7 @@ using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Enums;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions;
 
-namespace Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Extensions
+namespace Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Core.Extensions
 {
     /// <summary>
     /// This represents the extension entity for <see cref="SwaggerUI"/>.
