+++ src/Microsoft.Azure.Functions.Worker.Extensions.OpenApi/Microsoft.Azure.Functions.Worker.Extensions.OpenApi.csproj
@@ -10,8 +10,8 @@
     <Version>1.0.0</Version>
     <PackageId>Microsoft.Azure.Functions.Worker.Extensions.OpenApi</PackageId>
     <Title>Microsoft.Azure.Functions.Worker.Extensions.OpenApi</Title>
-    <Summary>This package helps render OpenAPI document and Swagger UI through Azure Functions endpoints through the out-of-process worker environment.</Summary>
-    <Description>This package helps render OpenAPI document and Swagger UI through Azure Functions endpoints through the out-of-process worker environment.</Description>
+    <Summary>This package helps render OpenAPI document and Swagger UI through Azure Functions endpoints through the out-of-process worker.</Summary>
+    <Description>This package helps render OpenAPI document and Swagger UI through Azure Functions endpoints through the out-of-process worker.</Description>
     <PackageTags>Azure-Functions, Open-API, Swagger, Swagger-UI</PackageTags>
     <AssemblyName>Microsoft.Azure.Functions.Worker.Extensions.OpenApi</AssemblyName>
     <RootNamespace>Microsoft.Azure.Functions.Worker.Extensions.OpenApi</RootNamespace>
@@ -26,20 +26,13 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Remove="Lists\**" />
-    <EmbeddedResource Remove="Lists\**" />
-    <None Remove="Lists\**" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.Azure.Core.NewtonsoftJson" Version="1.0.0" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.12" />
-    <!-- <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.0.3" OutputItemType="Analyzer" /> -->
-    <PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.1.0" />
+    <PackageReference Include="Microsoft.Azure.Functions.Worker.Core" Version="1.1.0" />
     <PackageReference Include="Microsoft.Azure.WebJobs.Script.Abstractions" Version="1.0.0-preview" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Core.Extensions\Microsoft.Azure.Functions.Worker.Extensions.OpenApi.Core.Extensions.csproj" />
     <ProjectReference Include="..\Microsoft.Azure.WebJobs.Extensions.OpenApi.Core\Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.csproj" />
   </ItemGroup>
