+++ src/Microsoft.Azure.WebJobs.Extensions.OpenApi.Core/Abstractions/IDocumentHelper.cs
@@ -15,41 +15,41 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions
     /// </summary>
     public interface IDocumentHelper
     {
-        /// <summary>
-        /// Gets the list of HTTP triggers.
-        /// </summary>
-        /// <param name="assembly">Assembly of Azure Function instance.</param>
-        /// <returns>List of <see cref="MethodInfo"/> instances representing HTTP triggers.</returns>
-        List<MethodInfo> GetHttpTriggerMethods(Assembly assembly);
+        // /// <summary>
+        // /// Gets the list of HTTP triggers.
+        // /// </summary>
+        // /// <param name="assembly">Assembly of Azure Function instance.</param>
+        // /// <returns>List of <see cref="MethodInfo"/> instances representing HTTP triggers.</returns>
+        // List<MethodInfo> GetHttpTriggerMethods(Assembly assembly);
 
-        /// <summary>
-        /// Gets the <see cref="FunctionNameAttribute"/> from the method.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <returns><see cref="FunctionNameAttribute"/> instance.</returns>
-        FunctionNameAttribute GetFunctionNameAttribute(MethodInfo element);
+        // /// <summary>
+        // /// Gets the <see cref="FunctionNameAttribute"/> from the method.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <returns><see cref="FunctionNameAttribute"/> instance.</returns>
+        // FunctionNameAttribute GetFunctionNameAttribute(MethodInfo element);
 
-        /// <summary>
-        /// Gets the <see cref="HttpTriggerAttribute"/> from the parameters of the method.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <returns><see cref="HttpTriggerAttribute"/> instance.</returns>
-        HttpTriggerAttribute GetHttpTriggerAttribute(MethodInfo element);
+        // /// <summary>
+        // /// Gets the <see cref="HttpTriggerAttribute"/> from the parameters of the method.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <returns><see cref="HttpTriggerAttribute"/> instance.</returns>
+        // HttpTriggerAttribute GetHttpTriggerAttribute(MethodInfo element);
 
-        /// <summary>
-        /// Gets the HTTP trigger endpoint.
-        /// </summary>
-        /// <param name="function"><see cref="FunctionNameAttribute"/> instance.</param>
-        /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
-        /// <returns>Function HTTP endpoint.</returns>
-        string GetHttpEndpoint(FunctionNameAttribute function, HttpTriggerAttribute trigger);
+        // /// <summary>
+        // /// Gets the HTTP trigger endpoint.
+        // /// </summary>
+        // /// <param name="function"><see cref="FunctionNameAttribute"/> instance.</param>
+        // /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
+        // /// <returns>Function HTTP endpoint.</returns>
+        // string GetHttpEndpoint(FunctionNameAttribute function, HttpTriggerAttribute trigger);
 
-        /// <summary>
-        /// Gets the HTTP verb.
-        /// </summary>
-        /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
-        /// <returns><see cref="OperationType"/> value.</returns>
-        OperationType GetHttpVerb(HttpTriggerAttribute trigger);
+        // /// <summary>
+        // /// Gets the HTTP verb.
+        // /// </summary>
+        // /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
+        // /// <returns><see cref="OperationType"/> value.</returns>
+        // OperationType GetHttpVerb(HttpTriggerAttribute trigger);
 
         /// <summary>
         /// Gets the <see cref="OpenApiPathItem"/> instance.
@@ -59,14 +59,14 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions
         /// <returns><see cref="OpenApiPathItem"/> instance.</returns>
         OpenApiPathItem GetOpenApiPath(string path, OpenApiPaths paths);
 
-        /// <summary>
-        /// Gets the <see cref="OpenApiOperation"/> instance.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <param name="function"><see cref="FunctionNameAttribute"/> instance.</param>
-        /// <param name="verb"><see cref="OperationType"/> value.</param>
-        /// <returns><see cref="OpenApiOperation"/> instance.</returns>
-        OpenApiOperation GetOpenApiOperation(MethodInfo element, FunctionNameAttribute function, OperationType verb);
+        // /// <summary>
+        // /// Gets the <see cref="OpenApiOperation"/> instance.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <param name="function"><see cref="FunctionNameAttribute"/> instance.</param>
+        // /// <param name="verb"><see cref="OperationType"/> value.</param>
+        // /// <returns><see cref="OpenApiOperation"/> instance.</returns>
+        // OpenApiOperation GetOpenApiOperation(MethodInfo element, FunctionNameAttribute function, OperationType verb);
 
         /// <summary>
         /// Gets the collection of <see cref="OpenApiSecurityRequirement"/> instances.
@@ -76,15 +76,15 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions
         /// <returns>Collection of <see cref="OpenApiSecurityRequirement"/> instance.</returns>
         List<OpenApiSecurityRequirement> GetOpenApiSecurityRequirement(MethodInfo element, NamingStrategy namingStrategy = null);
 
-        /// <summary>
-        /// Gets the list of <see cref="OpenApiParameter"/> instances.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
-        /// <param name="namingStrategy"><see cref="NamingStrategy"/> instance to create the JSON schema from .NET Types.</param>
-        /// <param name="collection"><see cref="VisitorCollection"/> instance to process parameters.</param>
-        /// <returns>List of <see cref="OpenApiParameter"/> instance.</returns>
-        List<OpenApiParameter> GetOpenApiParameters(MethodInfo element, HttpTriggerAttribute trigger, NamingStrategy namingStrategy, VisitorCollection collection);
+        // /// <summary>
+        // /// Gets the list of <see cref="OpenApiParameter"/> instances.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <param name="trigger"><see cref="HttpTriggerAttribute"/> instance.</param>
+        // /// <param name="namingStrategy"><see cref="NamingStrategy"/> instance to create the JSON schema from .NET Types.</param>
+        // /// <param name="collection"><see cref="VisitorCollection"/> instance to process parameters.</param>
+        // /// <returns>List of <see cref="OpenApiParameter"/> instance.</returns>
+        // List<OpenApiParameter> GetOpenApiParameters(MethodInfo element, HttpTriggerAttribute trigger, NamingStrategy namingStrategy, VisitorCollection collection);
 
         /// <summary>
         /// Gets the <see cref="OpenApiRequestBody"/> instance.
@@ -138,5 +138,12 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Abstractions
         /// <param name="namingStrategy"><see cref="NamingStrategy"/> instance to create the JSON schema from .NET Types.</param>
         /// <returns>Collection of <see cref="OpenApiSecurityScheme"/> instance.</returns>
         Dictionary<string, OpenApiSecurityScheme> GetOpenApiSecuritySchemes(List<MethodInfo> elements, NamingStrategy namingStrategy = null);
+
+        /// <summary>
+        /// Filters the route constraints.
+        /// </summary>
+        /// <param name="route">Route parameter value.</param>
+        /// <returns>Returns the route value without the constraints.</returns>
+        string FilterRouteConstraints(string route);
     }
 }
