+++ src/Microsoft.Azure.WebJobs.Extensions.OpenApi.Core/Extensions/MethodInfoExtensions.cs
@@ -1,9 +1,6 @@
-using System.Linq;
 using System.Reflection;
 
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
-
-using Microsoft.Azure.WebJobs;
 
 namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions
 {
@@ -12,41 +9,41 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions
     /// </summary>
     public static class MethodInfoExtensions
     {
+        // /// <summary>
+        // /// Gets the <see cref="FunctionNameAttribute"/> instance.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <returns><see cref="FunctionNameAttribute"/> instance.</returns>
+        // public static FunctionNameAttribute GetFunctionName(this MethodInfo element)
+        // {
+        //     element.ThrowIfNullOrDefault();
+
+        //     var function = element.GetCustomAttribute<FunctionNameAttribute>(inherit: false);
+
+        //     return function;
+        // }
+
+        // /// <summary>
+        // /// Gets the <see cref="HttpTriggerAttribute"/> instance.
+        // /// </summary>
+        // /// <param name="element"><see cref="MethodInfo"/> instance.</param>
+        // /// <returns><see cref="HttpTriggerAttribute"/> instance.</returns>
+        // public static HttpTriggerAttribute GetHttpTrigger(this MethodInfo element)
+        // {
+        //     element.ThrowIfNullOrDefault();
+
+        //     var trigger = element.GetParameters()
+        //                          .First()
+        //                          .GetCustomAttribute<HttpTriggerAttribute>(inherit: false);
+
+        //     return trigger;
+        // }
+
         /// <summary>
-        /// Gets the <see cref="FunctionNameAttribute"/> instance.
+        /// Gets the <see cref="OpenApiOperationAttribute"/> instance.
         /// </summary>
         /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <returns><see cref="FunctionNameAttribute"/> instance.</returns>
-        public static FunctionNameAttribute GetFunctionName(this MethodInfo element)
-        {
-            element.ThrowIfNullOrDefault();
-
-            var function = element.GetCustomAttribute<FunctionNameAttribute>(inherit: false);
-
-            return function;
-        }
-
-        /// <summary>
-        /// Gets the <see cref="HttpTriggerAttribute"/> instance.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <returns><see cref="HttpTriggerAttribute"/> instance.</returns>
-        public static HttpTriggerAttribute GetHttpTrigger(this MethodInfo element)
-        {
-            element.ThrowIfNullOrDefault();
-
-            var trigger = element.GetParameters()
-                                 .First()
-                                 .GetCustomAttribute<HttpTriggerAttribute>(inherit: false);
-
-            return trigger;
-        }
-
-        /// <summary>
-        /// Gets the <see cref="HttpTriggerAttribute"/> instance.
-        /// </summary>
-        /// <param name="element"><see cref="MethodInfo"/> instance.</param>
-        /// <returns><see cref="HttpTriggerAttribute"/> instance.</returns>
+        /// <returns><see cref="OpenApiOperationAttribute"/> instance.</returns>
         public static OpenApiOperationAttribute GetOpenApiOperation(this MethodInfo element)
         {
             element.ThrowIfNullOrDefault();
