+++ test/Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions.Tests/FunctionNameAttributeExtensionsTests.cs
@@ -3,20 +3,19 @@ using System.Reflection;
 
 using FluentAssertions;
 
+using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Tests.Fakes;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-using MethodInfoExtensions = Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Extensions.MethodInfoExtensions;
-
-namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Tests.Extensions
+namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Extensions.Tests
 {
     [TestClass]
-    public class MethodInfoExtensionsTests
+    public class FunctionNameAttributeExtensionsTests
     {
         [TestMethod]
         public void Given_Null_When_GetFunctionName_Invoked_Then_It_Should_Throw_Exception()
         {
-            Action action = () => MethodInfoExtensions.GetFunctionName(null);
+            Action action = () => FunctionNameAttributeExtensions.GetFunctionName(null);
 
             action.Should().Throw<ArgumentNullException>();
         }
@@ -27,18 +26,10 @@ namespace Microsoft.Azure.WebJobs.Extensions.OpenApi.Tests.Extensions
         {
             var method = typeof(FakeHttpTrigger).GetMethod(methodName, BindingFlags.Public | BindingFlags.Instance);
 
-            var result = MethodInfoExtensions.GetFunctionName(method);
+            var result = FunctionNameAttributeExtensions.GetFunctionName(method);
 
             result.Should().BeOfType<FunctionNameAttribute>();
             result.Name.Should().Be(expected);
-        }
-
-        [TestMethod]
-        public void Given_Null_When_GetHttpTrigger_Invoked_Then_It_Should_Throw_Exception()
-        {
-            Action action = () => MethodInfoExtensions.GetHttpTrigger(null);
-
-            action.Should().Throw<ArgumentNullException>();
         }
     }
 }
