+++ test/Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Tests.Fakes/Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Tests.Fakes.csproj
@@ -8,7 +8,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.1.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Http" Version="2.1.0" />
+    <PackageReference Include="Microsoft.Azure.WebJobs.Extensions.Http" Version="3.0.0" />
   </ItemGroup>
 
   <ItemGroup>
